move catalog-related files to 'storage/catalog'
