+++ src/storage/catalog/iceberg_catalog.cpp
@@ -1,5 +1,5 @@
-#include "storage/iceberg_schema_entry.hpp"
-#include "storage/iceberg_table_entry.hpp"
+#include "storage/catalog/iceberg_schema_entry.hpp"
+#include "storage/catalog/iceberg_table_entry.hpp"
 #include "storage/iceberg_transaction.hpp"
 #include "catalog_api.hpp"
 #include "catalog_utils.hpp"
@@ -13,7 +13,7 @@
 #include "duckdb/main/attached_database.hpp"
 #include "rest_catalog/objects/catalog_config.hpp"
 #include "duckdb/planner/operator/logical_create_table.hpp"
-#include "storage/iceberg_catalog.hpp"
+#include "storage/catalog/iceberg_catalog.hpp"
 #include "regex"
 #include "storage/iceberg_authorization.hpp"
 #include "storage/authorization/oauth2.hpp"
