+++ src/storage/catalog/iceberg_table_entry.cpp
@@ -1,6 +1,6 @@
-#include "storage/iceberg_catalog.hpp"
-#include "storage/iceberg_schema_entry.hpp"
-#include "storage/iceberg_table_entry.hpp"
+#include "storage/catalog/iceberg_catalog.hpp"
+#include "storage/catalog/iceberg_schema_entry.hpp"
+#include "storage/catalog/iceberg_table_entry.hpp"
 #include "duckdb/storage/statistics/base_statistics.hpp"
 #include "duckdb/storage/table_storage_info.hpp"
 #include "duckdb/main/extension/extension_loader.hpp"
