+++ src/storage/catalog/iceberg_table_set.cpp
@@ -2,15 +2,15 @@
 #include "catalog_utils.hpp"
 #include "iceberg_logging.hpp"
 
-#include "storage/iceberg_catalog.hpp"
-#include "storage/iceberg_table_set.hpp"
+#include "storage/catalog/iceberg_catalog.hpp"
+#include "storage/catalog/iceberg_table_set.hpp"
 
-#include "storage/iceberg_table_entry.hpp"
+#include "storage/catalog/iceberg_table_entry.hpp"
 #include "storage/iceberg_transaction.hpp"
 #include "storage/authorization/sigv4.hpp"
 #include "storage/iceberg_table_information.hpp"
 #include "storage/authorization/oauth2.hpp"
-#include "storage/iceberg_schema_entry.hpp"
+#include "storage/catalog/iceberg_schema_entry.hpp"
 #include "metadata/iceberg_partition_spec.hpp"
 
 #include "duckdb/parser/expression/constant_expression.hpp"
