+++ src/storage/iceberg_table_information.cpp
@@ -7,8 +7,8 @@
 #include "duckdb/common/types/timestamp.hpp"
 #include "storage/iceberg_transaction.hpp"
 #include "storage/iceberg_transaction_data.hpp"
-#include "storage/iceberg_schema_entry.hpp"
-#include "storage/iceberg_catalog.hpp"
+#include "storage/catalog/iceberg_schema_entry.hpp"
+#include "storage/catalog/iceberg_catalog.hpp"
 #include "storage/iceberg_authorization.hpp"
 #include "storage/authorization/oauth2.hpp"
 #include "storage/authorization/sigv4.hpp"
