+++ src/storage/iceberg_update.cpp
@@ -2,8 +2,8 @@
 #include "storage/iceberg_delete.hpp"
 #include "storage/iceberg_insert.hpp"
 #include "storage/iceberg_transaction.hpp"
-#include "storage/iceberg_catalog.hpp"
-#include "storage/iceberg_table_entry.hpp"
+#include "storage/catalog/iceberg_catalog.hpp"
+#include "storage/catalog/iceberg_table_entry.hpp"
 #include "storage/iceberg_table_information.hpp"
 #include "duckdb/planner/operator/logical_update.hpp"
 #include "duckdb/parallel/thread_context.hpp"
