+++ src/storage/table_update/iceberg_add_snapshot.cpp
@@ -1,6 +1,6 @@
 #include "storage/table_update/iceberg_add_snapshot.hpp"
 #include "metadata/iceberg_manifest_list.hpp"
-#include "storage/iceberg_table_set.hpp"
+#include "storage/catalog/iceberg_table_set.hpp"
 
 #include "duckdb/common/enums/catalog_type.hpp"
 #include "duckdb/catalog/catalog_entry/copy_function_catalog_entry.hpp"
